[AssetCondition] Refactor the testing suite
