+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/asset_daemon_scenario.py
@@ -1,44 +1,27 @@
+import dataclasses
 import datetime
 import itertools
-import json
 import logging
-import os
-import sys
-from collections import namedtuple
 from concurrent.futures import ThreadPoolExecutor
-from contextlib import contextmanager
+from dataclasses import dataclass, field
 from typing import (
     Any,
     Callable,
-    Iterable,
     NamedTuple,
     Optional,
     Sequence,
     Tuple,
     Type,
-    Union,
     cast,
 )
 
 import dagster._check as check
-import mock
-import pendulum
 from dagster import (
-    AssetExecutionContext,
     AssetKey,
-    AssetsDefinition,
-    AssetSpec,
-    AutoMaterializePolicy,
     DagsterInstance,
-    DagsterRunStatus,
-    Definitions,
     MultiPartitionKey,
-    RepositoryDefinition,
     RunRequest,
     RunsFilter,
-    asset,
-    create_repository_using_definitions_args,
-    materialize,
 )
 from dagster._core.definitions.asset_condition.asset_condition import (
     AssetConditionEvaluation,
@@ -57,25 +40,13 @@ from dagster._core.definitions.auto_materialize_rule import AutoMaterializeRule
 from dagster._core.definitions.auto_materialize_rule_evaluation import (
     AutoMaterializeRuleEvaluationData,
 )
-from dagster._core.definitions.auto_materialize_sensor_definition import (
-    AutoMaterializeSensorDefinition,
-)
-from dagster._core.definitions.decorators.asset_decorator import multi_asset
 from dagster._core.definitions.events import AssetKeyPartitionKey, CoercibleToAssetKey
-from dagster._core.definitions.executor_definition import in_process_executor
-from dagster._core.definitions.internal_asset_graph import InternalAssetGraph
-from dagster._core.definitions.partition import PartitionsDefinition
-from dagster._core.definitions.repository_definition.repository_definition import (
-    PendingRepositoryDefinition,
-)
 from dagster._core.definitions.repository_definition.valid_definitions import (
     SINGLETON_REPOSITORY_NAME,
 )
-from dagster._core.host_representation.external_data import external_repository_data_from_def
 from dagster._core.host_representation.origin import (
     ExternalInstigatorOrigin,
     ExternalRepositoryOrigin,
-    InProcessCodeLocationOrigin,
 )
 from dagster._core.scheduler.instigation import (
     SensorInstigatorData,
@@ -83,11 +54,8 @@ from dagster._core.scheduler.instigation import (
 )
 from dagster._core.storage.tags import PARTITION_NAME_TAG
 from dagster._core.test_utils import (
-    InProcessTestWorkspaceLoadTarget,
-    create_test_daemon_workspace_context,
     wait_for_futures,
 )
-from dagster._core.types.loadable_target_origin import LoadableTargetOrigin
 from dagster._daemon.asset_daemon import (
     _PRE_SENSOR_AUTO_MATERIALIZE_ORIGIN_ID,
     _PRE_SENSOR_AUTO_MATERIALIZE_SELECTOR_ID,
@@ -96,9 +64,6 @@ from dagster._daemon.asset_daemon import (
     asset_daemon_cursor_from_instigator_serialized_cursor,
     get_current_evaluation_id,
 )
-from dagster._serdes import (
-    create_snapshot_id,
-)
 from dagster._serdes.serdes import (
     DeserializationError,
     deserialize_value,
@@ -106,48 +71,12 @@ from dagster._serdes.serdes import (
 )
 from dagster._seven.compat.pendulum import pendulum_freeze_time
 
-from .base_scenario import FAIL_TAG, run_request
-
-
-def get_code_location_origin(
-    scenario_state: "AssetDaemonScenarioState",
-    location_name="test_location",
-    repository_name=SINGLETON_REPOSITORY_NAME,
-) -> InProcessCodeLocationOrigin:
-    """Hacky method to allow us to point a code location at a module-scoped attribute, even though
-    the attribute is not defined until the scenario is run.
-    """
-    repository = create_repository_using_definitions_args(
-        name=repository_name,
-        assets=scenario_state.assets,
-        executor=in_process_executor,
-        sensors=scenario_state.auto_materialize_sensors,
-    )
-
-    return _get_code_location_origin_from_repository(
-        cast(RepositoryDefinition, repository), location_name=location_name
-    )
-
-
-def _get_code_location_origin_from_repository(repository: RepositoryDefinition, location_name: str):
-    attribute_name = (
-        f"_asset_daemon_target_{create_snapshot_id(external_repository_data_from_def(repository))}"
-    )
-
-    if attribute_name not in globals():
-        globals()[attribute_name] = repository
-
-    return InProcessCodeLocationOrigin(
-        loadable_target_origin=LoadableTargetOrigin(
-            executable_path=sys.executable,
-            module_name=(
-                "dagster_tests.definitions_tests.auto_materialize_tests.asset_daemon_scenario"
-            ),
-            working_directory=os.getcwd(),
-            attribute=attribute_name,
-        ),
-        location_name=location_name,
-    )
+from .base_scenario import run_request
+from .scenario_state import (
+    ScenarioSpec,
+    ScenarioState,
+    get_code_location_origin,
+)
 
 
 def day_partition_key(time: datetime.datetime, delta: int = 0) -> str:
@@ -214,22 +143,8 @@ class AssetRuleEvaluationSpec(NamedTuple):
         return AssetSubsetWithMetadata(subset=subset, metadata=metadata)
 
 
-class AssetSpecWithPartitionsDef(
-    namedtuple(
-        "AssetSpecWithPartitionsDef",
-        AssetSpec._fields + ("partitions_def",),
-        defaults=(None,) * (1 + len(AssetSpec._fields)),
-    )
-): ...
-
-
-class MultiAssetSpec(NamedTuple):
-    specs: Sequence[AssetSpec]
-    partitions_def: Optional[PartitionsDefinition] = None
-    can_subset: bool = False
-
-
-class AssetDaemonScenarioState(NamedTuple):
+@dataclass(frozen=True)
+class AssetDaemonScenarioState(ScenarioState):
     """Specifies the state of a given AssetDaemonScenario. This state can be modified by changing
     the set of asset definitions it contains, executing runs, updating the time, evaluating ticks, etc.
 
@@ -243,183 +158,19 @@ class AssetDaemonScenarioState(NamedTuple):
         current_time (datetime): The current time of the scenario.
     """
 
-    asset_specs: Sequence[Union[AssetSpec, AssetSpecWithPartitionsDef, MultiAssetSpec]]
-    additional_repositories: Sequence[Union[RepositoryDefinition, PendingRepositoryDefinition]] = []
-
-    current_time: datetime.datetime = pendulum.now("UTC")
-    run_requests: Sequence[RunRequest] = []
-    serialized_cursor: str = serialize_value(AssetDaemonCursor.empty(0))
-    evaluations: Sequence[AssetConditionEvaluation] = []
-    logger: logging.Logger = logging.getLogger("dagster.amp")
+    run_requests: Sequence[RunRequest] = field(default_factory=list)
+    serialized_cursor: str = field(default=serialize_value(AssetDaemonCursor.empty(0)))
+    evaluations: Sequence[AssetConditionEvaluation] = field(default_factory=list)
+    logger: logging.Logger = field(default=logging.getLogger("dagster.amp"))
     tick_index: int = 1
-    # this is set by the scenario runner
-    scenario_instance: Optional[DagsterInstance] = None
+
+    # set by scenario runner
     is_daemon: bool = False
     sensor_name: Optional[str] = None
-    auto_materialize_sensors: Optional[Sequence[AutoMaterializeSensorDefinition]] = None
     threadpool_executor: Optional[ThreadPoolExecutor] = None
 
-    @property
-    def instance(self) -> DagsterInstance:
-        return check.not_none(self.scenario_instance)
-
-    @property
-    def assets(self) -> Sequence[AssetsDefinition]:
-        def compute_fn(context: AssetExecutionContext) -> None:
-            fail_keys = {
-                AssetKey.from_coercible(s)
-                for s in json.loads(context.run.tags.get(FAIL_TAG) or "[]")
-            }
-            for asset_key in context.selected_asset_keys:
-                if asset_key in fail_keys:
-                    raise Exception("Asset failed")
-
-        assets = []
-        for spec in self.asset_specs:
-            if isinstance(spec, MultiAssetSpec):
-
-                @multi_asset(**spec._asdict())
-                def _multi_asset(context: AssetExecutionContext):
-                    return compute_fn(context)
-
-                assets.append(_multi_asset)
-            else:
-                params = {
-                    "key",
-                    "deps",
-                    "group_name",
-                    "code_version",
-                    "auto_materialize_policy",
-                    "freshness_policy",
-                    "partitions_def",
-                }
-                assets.append(
-                    asset(
-                        compute_fn=compute_fn,
-                        **{k: v for k, v in spec._asdict().items() if k in params},
-                    )
-                )
-        return assets
-
-    @property
-    def defs(self) -> Definitions:
-        return Definitions(assets=self.assets, sensors=self.auto_materialize_sensors)
-
-    @property
-    def asset_graph(self) -> AssetGraph:
-        return InternalAssetGraph.from_assets(self.assets)
-
-    def with_additional_repositories(
-        self,
-        additional_repositories: Sequence[Union[RepositoryDefinition, PendingRepositoryDefinition]],
-    ) -> "AssetDaemonScenarioState":
-        return self._replace(additional_repositories=additional_repositories)
-
-    def with_asset_properties(
-        self, keys: Optional[Iterable[CoercibleToAssetKey]] = None, **kwargs
-    ) -> "AssetDaemonScenarioState":
-        """Convenience method to update the properties of one or more assets in the scenario state."""
-        new_asset_specs = []
-        for spec in self.asset_specs:
-            if isinstance(spec, MultiAssetSpec):
-                partitions_def = kwargs.get("partitions_def", spec.partitions_def)
-                new_multi_specs = [
-                    s._replace(**{k: v for k, v in kwargs.items() if k != "partitions_def"})
-                    if keys is None or s.key in keys
-                    else s
-                    for s in spec.specs
-                ]
-                new_asset_specs.append(
-                    spec._replace(partitions_def=partitions_def, specs=new_multi_specs)
-                )
-            else:
-                if keys is None or spec.key in {AssetKey.from_coercible(key) for key in keys}:
-                    if "partitions_def" in kwargs:
-                        # partitions_def is not a field on AssetSpec, so we need to do this hack
-                        new_asset_specs.append(
-                            AssetSpecWithPartitionsDef(**{**spec._asdict(), **kwargs})
-                        )
-                    else:
-                        new_asset_specs.append(spec._replace(**kwargs))
-                else:
-                    new_asset_specs.append(spec)
-        return self._replace(asset_specs=new_asset_specs)
-
-    def with_auto_materialize_sensors(
-        self,
-        sensors: Optional[Sequence[AutoMaterializeSensorDefinition]],
-    ):
-        return self._replace(auto_materialize_sensors=sensors)
-
     def with_serialized_cursor(self, serialized_cursor: str) -> "AssetDaemonScenarioState":
-        return self._replace(serialized_cursor=serialized_cursor)
-
-    def with_all_eager(
-        self, max_materializations_per_minute: int = 1
-    ) -> "AssetDaemonScenarioState":
-        return self.with_asset_properties(
-            auto_materialize_policy=AutoMaterializePolicy.eager(
-                max_materializations_per_minute=max_materializations_per_minute
-            )
-        )
-
-    def with_current_time(self, time: str) -> "AssetDaemonScenarioState":
-        return self._replace(current_time=pendulum.parse(time))
-
-    def with_current_time_advanced(self, **kwargs) -> "AssetDaemonScenarioState":
-        # hacky support for adding years
-        if "years" in kwargs:
-            kwargs["days"] = kwargs.get("days", 0) + 365 * kwargs.pop("years")
-        return self._replace(current_time=self.current_time + datetime.timedelta(**kwargs))
-
-    def with_runs(self, *run_requests: RunRequest) -> "AssetDaemonScenarioState":
-        start = datetime.datetime.now()
-
-        def test_time_fn() -> float:
-            # this function will increment the current timestamp in real time, relative to the
-            # fake current_time on the scenario state
-            return (self.current_time + (datetime.datetime.now() - start)).timestamp()
-
-        with pendulum_freeze_time(self.current_time), mock.patch("time.time", new=test_time_fn):
-            for rr in run_requests:
-                materialize(
-                    assets=self.assets,
-                    instance=self.instance,
-                    partition_key=rr.partition_key,
-                    tags=rr.tags,
-                    raise_on_error=False,
-                    selection=rr.asset_selection,
-                )
-        # increment current_time by however much time elapsed during the materialize call
-        return self._replace(current_time=pendulum.from_timestamp(test_time_fn()))
-
-    def with_not_started_runs(self) -> "AssetDaemonScenarioState":
-        """Execute all runs in the NOT_STARTED state and delete them from the instance. The scenario
-        adds in the run requests from previous ticks as runs in the NOT_STARTED state, so this method
-        executes requested runs from previous ticks.
-        """
-        not_started_runs = self.instance.get_runs(
-            filters=RunsFilter(statuses=[DagsterRunStatus.NOT_STARTED])
-        )
-        for run in not_started_runs:
-            self.instance.delete_run(run_id=run.run_id)
-        return self.with_runs(
-            *[
-                run_request(
-                    asset_keys=list(run.asset_selection or set()),
-                    partition_key=run.tags.get(PARTITION_NAME_TAG),
-                )
-                for run in not_started_runs
-            ]
-        )
-
-    def with_dynamic_partitions(
-        self, partitions_def_name: str, partition_keys: Sequence[str]
-    ) -> "AssetDaemonScenarioState":
-        self.instance.add_dynamic_partitions(
-            partitions_def_name=partitions_def_name, partition_keys=partition_keys
-        )
-        return self
+        return dataclasses.replace(self, serialized_cursor=serialized_cursor)
 
     def _evaluate_tick_fast(
         self,
@@ -455,48 +206,13 @@ class AssetDaemonScenarioState(NamedTuple):
         # make sure these run requests are available on the instance
         for request in new_run_requests:
             asset_selection = check.not_none(request.asset_selection)
-            job_def = self.defs.get_implicit_job_def_for_assets(asset_selection)
+            job_def = self.scenario_spec.defs.get_implicit_job_def_for_assets(asset_selection)
             self.instance.create_run_for_job(
                 job_def=check.not_none(job_def),
                 asset_selection=set(asset_selection),
                 tags=request.tags,
             )
         return new_run_requests, new_cursor, new_evaluations
-
-    @contextmanager
-    def _create_workspace_context(self):
-        origins = [
-            _get_code_location_origin_from_repository(repository, f"extra_location_{i}")
-            for i, repository in enumerate(self.additional_repositories)
-        ] + [
-            get_code_location_origin(self),
-        ]
-
-        target = InProcessTestWorkspaceLoadTarget(origins)
-        with create_test_daemon_workspace_context(
-            workspace_load_target=target, instance=self.instance
-        ) as workspace_context:
-            yield workspace_context
-
-    @contextmanager
-    def _get_external_sensor(self, sensor_name):
-        with self._create_workspace_context() as workspace_context:
-            workspace = workspace_context.create_request_context()
-            sensor = next(
-                iter(workspace.get_code_location("test_location").get_repositories().values())
-            ).get_external_sensor(sensor_name)
-            assert sensor
-            yield sensor
-
-    def start_sensor(self, sensor_name: str):
-        with self._get_external_sensor(sensor_name) as sensor:
-            self.instance.start_sensor(sensor)
-        return self
-
-    def stop_sensor(self, sensor_name: str):
-        with self._get_external_sensor(sensor_name) as sensor:
-            self.instance.stop_sensor(sensor.get_external_origin_id(), sensor.selector_id, sensor)
-        return self
 
     def _evaluate_tick_daemon(
         self,
@@ -592,7 +308,8 @@ class AssetDaemonScenarioState(NamedTuple):
             else:
                 new_run_requests, new_cursor, new_evaluations = self._evaluate_tick_fast()
 
-        return self._replace(
+        return dataclasses.replace(
+            self,
             run_requests=new_run_requests,
             serialized_cursor=serialize_value(new_cursor),
             evaluations=new_evaluations,
@@ -608,7 +325,7 @@ class AssetDaemonScenarioState(NamedTuple):
     def get_sensor_origin(self):
         if not self.sensor_name:
             return None
-        code_location_origin = get_code_location_origin(self)
+        code_location_origin = get_code_location_origin(self.scenario_spec)
         return ExternalInstigatorOrigin(
             external_repository_origin=ExternalRepositoryOrigin(
                 code_location_origin=code_location_origin,
@@ -809,22 +526,19 @@ class AssetDaemonScenario(NamedTuple):
     """
 
     id: str
-    initial_state: AssetDaemonScenarioState
+    initial_spec: ScenarioSpec
     execution_fn: Callable[[AssetDaemonScenarioState], AssetDaemonScenarioState]
 
     def evaluate_fast(self) -> None:
-        self.initial_state.logger.setLevel(logging.DEBUG)
-        self.execution_fn(
-            self.initial_state._replace(scenario_instance=DagsterInstance.ephemeral())
-        )
+        self.execution_fn(AssetDaemonScenarioState(self.initial_spec))
 
     def evaluate_daemon(
         self, instance: DagsterInstance, sensor_name: Optional[str] = None, threadpool_executor=None
     ) -> "AssetDaemonScenarioState":
-        self.initial_state.logger.setLevel(logging.DEBUG)
         return self.execution_fn(
-            self.initial_state._replace(
-                scenario_instance=instance,
+            AssetDaemonScenarioState(
+                self.initial_spec,
+                instance=instance,
                 is_daemon=True,
                 sensor_name=sensor_name,
                 threadpool_executor=threadpool_executor,
