+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/test_asset_daemon.py
@@ -9,7 +9,6 @@ from dagster import (
     AssetSpec,
     AutoMaterializeRule,
     DagsterInstance,
-    create_repository_using_definitions_args,
     instance_for_test,
 )
 from dagster._core.definitions.asset_daemon_cursor import AssetDaemonCursor
@@ -18,7 +17,6 @@ from dagster._core.definitions.auto_materialize_policy import AutoMaterializePol
 from dagster._core.definitions.auto_materialize_sensor_definition import (
     AutoMaterializeSensorDefinition,
 )
-from dagster._core.definitions.executor_definition import in_process_executor
 from dagster._core.definitions.sensor_definition import DefaultSensorStatus
 from dagster._core.scheduler.instigation import (
     InstigatorStatus,
@@ -46,13 +44,13 @@ from dagster._daemon.asset_daemon import (
 )
 from dagster._serdes.serdes import serialize_value
 
+from dagster_tests.definitions_tests.auto_materialize_tests.scenario_state import ScenarioSpec
+
 from .asset_daemon_scenario import (
     AssetDaemonScenario,
-    AssetDaemonScenarioState,
     AssetRuleEvaluationSpec,
 )
 from .base_scenario import (
-    asset_def,
     run_request,
 )
 from .updated_scenarios.asset_daemon_scenario_states import (
@@ -100,38 +98,32 @@ def _get_threadpool_executor(instance: DagsterInstance):
 # just run over a subset of the total scenarios
 daemon_scenarios = [*basic_scenarios, *partition_scenarios]
 
-extra_repo_assets = [
-    asset_def("extra_asset1", auto_materialize_policy=AutoMaterializePolicy.eager()),
-    asset_def(
-        "extra_asset2", ["extra_asset1"], auto_materialize_policy=AutoMaterializePolicy.eager()
-    ),
-]
 
 # Additional repo with assets that should be not be included in the evaluation
-extra_repo = create_repository_using_definitions_args(
-    name="extra_repository",
-    assets=extra_repo_assets,
-    executor=in_process_executor,
+extra_definitions = ScenarioSpec(
+    [
+        AssetSpec("extra_asset1", auto_materialize_policy=AutoMaterializePolicy.eager()),
+        AssetSpec(
+            "extra_asset2",
+            deps=["extra_asset1"],
+            auto_materialize_policy=AutoMaterializePolicy.eager(),
+        ),
+    ]
 )
 
-first_asset_in_other_repo = [
-    asset_def("asset1", auto_materialize_policy=AutoMaterializePolicy.eager())
-]
-
-other_repo = create_repository_using_definitions_args(
-    name="extra_repository",
-    assets=first_asset_in_other_repo,
-    executor=in_process_executor,
+other_repo_definitions = ScenarioSpec(
+    asset_specs=[AssetSpec("asset1", auto_materialize_policy=AutoMaterializePolicy.eager())]
 )
 
-second_asset_in_our_repo = AssetDaemonScenarioState(
-    asset_specs=[AssetSpec("asset2", deps=["asset1"])],
-    additional_repositories=[other_repo],
-)
+
+second_asset_in_our_repo = ScenarioSpec(
+    asset_specs=[AssetSpec("asset2", deps=["asset1"])]
+).with_additional_repositories([other_repo_definitions])
+
 
 cross_repo_sensor_scenario = AssetDaemonScenario(
     id="two_cross_repo_assets",
-    initial_state=second_asset_in_our_repo.with_all_eager(),
+    initial_spec=second_asset_in_our_repo.with_all_eager(),
     # No runs since the first asset is not yet materialized and is not being considered by this sensor
     execution_fn=lambda state: state.evaluate_tick().assert_requested_runs(),
 )
@@ -141,11 +133,11 @@ auto_materialize_sensor_scenarios = [
     cross_repo_sensor_scenario,
     AssetDaemonScenario(
         id="basic_hourly_cron_unpartitioned",
-        initial_state=one_asset.with_asset_properties(
+        initial_spec=one_asset.with_asset_properties(
             auto_materialize_policy=get_cron_policy(basic_hourly_cron_schedule)
         )
         .with_current_time("2020-01-01T00:05")
-        .with_additional_repositories([extra_repo]),
+        .with_additional_repositories([extra_definitions]),
         execution_fn=lambda state: state.evaluate_tick()
         .assert_requested_runs(run_request(["A"]))
         .assert_evaluation("A", [AssetRuleEvaluationSpec(basic_hourly_cron_rule)])
@@ -165,8 +157,8 @@ auto_materialize_sensor_scenarios = [
     ),
     AssetDaemonScenario(
         id="sensor_interval_respected",
-        initial_state=two_assets_in_sequence.with_all_eager().with_additional_repositories(
-            [extra_repo]
+        initial_spec=two_assets_in_sequence.with_all_eager().with_additional_repositories(
+            [extra_definitions]
         ),
         execution_fn=lambda state: state.with_runs(run_request(["A", "B"]))
         .evaluate_tick()
@@ -183,7 +175,7 @@ auto_materialize_sensor_scenarios = [
     ),
     AssetDaemonScenario(
         id="one_asset_never_materialized",
-        initial_state=one_asset.with_all_eager().with_additional_repositories([extra_repo]),
+        initial_spec=one_asset.with_all_eager().with_additional_repositories([extra_definitions]),
         execution_fn=lambda state: state.evaluate_tick()
         .assert_requested_runs(run_request(asset_keys=["A"]))
         .assert_evaluation(
@@ -192,7 +184,7 @@ auto_materialize_sensor_scenarios = [
     ),
     AssetDaemonScenario(
         id="one_asset_already_launched",
-        initial_state=one_asset.with_all_eager().with_additional_repositories([extra_repo]),
+        initial_spec=one_asset.with_all_eager().with_additional_repositories([extra_definitions]),
         execution_fn=lambda state: state.evaluate_tick()
         .assert_requested_runs(run_request(asset_keys=["A"]))
         .with_current_time_advanced(seconds=30)
@@ -280,7 +272,7 @@ def _create_tick(instance: DagsterInstance, status: TickStatus, timestamp: float
 # valid at the start may not be valid when repeated.
 daemon_scenario = AssetDaemonScenario(
     id="simple_daemon_scenario",
-    initial_state=two_assets_in_sequence.with_asset_properties(
+    initial_spec=two_assets_in_sequence.with_asset_properties(
         partitions_def=two_partitions_def
     ).with_all_eager(2),
     execution_fn=lambda state: state.evaluate_tick(),
@@ -299,7 +291,7 @@ def test_daemon_paused() -> None:
 
         set_auto_materialize_paused(instance, False)
 
-        state = daemon_scenario._replace(initial_state=state).evaluate_daemon(instance)
+        state = daemon_scenario.execution_fn(state)
         ticks = _get_asset_daemon_ticks(instance)
 
         assert len(ticks) == 1
@@ -309,7 +301,7 @@ def test_daemon_paused() -> None:
         assert ticks[0].tick_data.end_timestamp == state.current_time.timestamp()
         assert ticks[0].tick_data.auto_materialize_evaluation_id == 1
 
-        state = daemon_scenario._replace(initial_state=state).evaluate_daemon(instance)
+        state = daemon_scenario.execution_fn(state)
         ticks = _get_asset_daemon_ticks(instance)
 
         # no new runs, so tick is now skipped
@@ -320,13 +312,11 @@ def test_daemon_paused() -> None:
         assert ticks[-1].tick_data.auto_materialize_evaluation_id == 2
 
 
-three_assets = AssetDaemonScenarioState(
-    asset_specs=[AssetSpec("A"), AssetSpec("B"), AssetSpec("C")]
-)
+three_assets = ScenarioSpec(asset_specs=[AssetSpec("A"), AssetSpec("B"), AssetSpec("C")])
 
 daemon_sensor_scenario = AssetDaemonScenario(
     id="simple_daemon_scenario",
-    initial_state=three_assets.with_auto_materialize_sensors(
+    initial_spec=three_assets.with_sensors(
         [
             AutoMaterializeSensorDefinition(
                 name="auto_materialize_sensor_a",
@@ -730,7 +720,7 @@ def test_auto_materialize_sensor_ticks(num_threads):
 
 def test_default_purge() -> None:
     with get_daemon_instance() as instance:
-        scenario_time = daemon_scenario.initial_state.current_time
+        scenario_time = daemon_scenario.initial_spec.current_time
         _create_tick(
             instance, TickStatus.SKIPPED, (scenario_time - datetime.timedelta(days=8)).timestamp()
         )
