+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/updated_scenarios/asset_daemon_scenario_states.py
@@ -10,7 +10,7 @@ from dagster._core.definitions.time_window_partitions import (
     HourlyPartitionsDefinition,
 )
 
-from ..asset_daemon_scenario import AssetDaemonScenarioState, MultiAssetSpec
+from ..scenario_state import MultiAssetSpec, ScenarioSpec
 
 ############
 # PARTITIONS
@@ -36,25 +36,25 @@ self_partition_mapping = TimeWindowPartitionMapping(start_offset=-1, end_offset=
 ##############
 # BASIC STATES
 ##############
-one_asset = AssetDaemonScenarioState(asset_specs=[AssetSpec("A")])
+one_asset = ScenarioSpec(asset_specs=[AssetSpec("A")])
 
-two_assets_in_sequence = AssetDaemonScenarioState(
+two_assets_in_sequence = ScenarioSpec(
     asset_specs=[AssetSpec("A"), AssetSpec("B", deps=["A"])],
 )
 
-three_assets_in_sequence = AssetDaemonScenarioState(
+three_assets_in_sequence = ScenarioSpec(
     asset_specs=[AssetSpec("A"), AssetSpec("B", deps=["A"]), AssetSpec("C", deps=["B"])],
 )
 
-two_assets_depend_on_one = AssetDaemonScenarioState(
+two_assets_depend_on_one = ScenarioSpec(
     asset_specs=[AssetSpec("A"), AssetSpec("B", deps=["A"]), AssetSpec("C", deps=["A"])]
 )
 
-one_asset_depends_on_two = AssetDaemonScenarioState(
+one_asset_depends_on_two = ScenarioSpec(
     asset_specs=[AssetSpec("A"), AssetSpec("B"), AssetSpec("C", deps=["A", "B"])]
 )
 
-diamond = AssetDaemonScenarioState(
+diamond = ScenarioSpec(
     asset_specs=[
         AssetSpec(key="A"),
         AssetSpec(key="B", deps=["A"]),
@@ -63,12 +63,8 @@ diamond = AssetDaemonScenarioState(
     ]
 )
 
-three_assets_not_subsettable = AssetDaemonScenarioState(
-    asset_specs=[
-        MultiAssetSpec(
-            specs=[AssetSpec("A"), AssetSpec("B"), AssetSpec("C")],
-        )
-    ]
+three_assets_not_subsettable = ScenarioSpec(
+    asset_specs=[MultiAssetSpec(specs=[AssetSpec("A"), AssetSpec("B"), AssetSpec("C")])]
 )
 
 ##################
