+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/updated_scenarios/basic_scenarios.py
@@ -5,10 +5,10 @@ from dagster._core.definitions.auto_materialize_rule_evaluation import (
 
 from ..asset_daemon_scenario import (
     AssetDaemonScenario,
-    AssetDaemonScenarioState,
     AssetRuleEvaluationSpec,
 )
 from ..base_scenario import run_request
+from ..scenario_state import ScenarioSpec
 from .asset_daemon_scenario_states import (
     diamond,
     one_asset,
@@ -21,7 +21,7 @@ from .asset_daemon_scenario_states import (
 basic_scenarios = [
     AssetDaemonScenario(
         id="one_asset_never_materialized",
-        initial_state=one_asset.with_all_eager(),
+        initial_spec=one_asset.with_all_eager(),
         execution_fn=lambda state: state.evaluate_tick()
         .assert_requested_runs(run_request(asset_keys=["A"]))
         .assert_evaluation(
@@ -30,7 +30,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="one_asset_already_launched",
-        initial_state=one_asset.with_all_eager(),
+        initial_spec=one_asset.with_all_eager(),
         execution_fn=lambda state: state.evaluate_tick()
         .assert_requested_runs(run_request(asset_keys=["A"]))
         .evaluate_tick()
@@ -38,7 +38,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="two_assets_in_sequence_never_materialized",
-        initial_state=two_assets_in_sequence.with_all_eager(),
+        initial_spec=two_assets_in_sequence.with_all_eager(),
         execution_fn=lambda state: state.evaluate_tick()
         .assert_requested_runs(run_request(asset_keys=["A", "B"]))
         .assert_evaluation(
@@ -60,7 +60,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="parent_materialized_child_not",
-        initial_state=two_assets_in_sequence.with_all_eager(),
+        initial_spec=two_assets_in_sequence.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A"]))
         .evaluate_tick()
         .assert_requested_runs(run_request(["B"]))
@@ -82,7 +82,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="parent_materialized_launch_two_children",
-        initial_state=two_assets_depend_on_one.with_all_eager(),
+        initial_spec=two_assets_depend_on_one.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A"]))
         .evaluate_tick()
         .assert_requested_runs(run_request(["B", "C"]))
@@ -115,7 +115,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="parent_materialized_with_source_asset_launch_child",
-        initial_state=AssetDaemonScenarioState(
+        initial_spec=ScenarioSpec(
             asset_specs=[AssetSpec("A"), AssetSpec("B", deps=["A", "source"])]
         ).with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A"]))
@@ -124,7 +124,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="parent_rematerialized_after_tick",
-        initial_state=two_assets_in_sequence.with_all_eager(),
+        initial_spec=two_assets_in_sequence.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A", "B"]))
         .evaluate_tick()
         .assert_requested_runs()
@@ -137,14 +137,14 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="parent_rematerialized",
-        initial_state=two_assets_in_sequence.with_all_eager(),
+        initial_spec=two_assets_in_sequence.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A", "B"]), run_request(["A"]))
         .evaluate_tick()
         .assert_requested_runs(run_request(["B"])),
     ),
     AssetDaemonScenario(
         id="one_parent_materialized_other_never_materialized",
-        initial_state=one_asset_depends_on_two.with_all_eager(),
+        initial_spec=one_asset_depends_on_two.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A"]))
         .evaluate_tick()
         .assert_requested_runs(run_request(["B", "C"]))
@@ -161,7 +161,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="one_parent_materialized_others_materialized_before",
-        initial_state=one_asset_depends_on_two.with_all_eager(),
+        initial_spec=one_asset_depends_on_two.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A", "B", "C"]))
         .evaluate_tick()
         .assert_requested_runs()
@@ -183,21 +183,21 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="diamond_never_materialized",
-        initial_state=diamond.with_all_eager(),
+        initial_spec=diamond.with_all_eager(),
         execution_fn=lambda state: state.evaluate_tick().assert_requested_runs(
             run_request(["A", "B", "C", "D"])
         ),
     ),
     AssetDaemonScenario(
         id="diamond_only_root_materialized",
-        initial_state=diamond.with_all_eager(),
+        initial_spec=diamond.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A"]))
         .evaluate_tick()
         .assert_requested_runs(run_request(["B", "C", "D"])),
     ),
     AssetDaemonScenario(
         id="diamond_root_rematerialized",
-        initial_state=diamond.with_all_eager(),
+        initial_spec=diamond.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A", "B", "C", "D"]))
         .evaluate_tick()
         .assert_requested_runs()
@@ -243,7 +243,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="diamond_root_and_one_in_middle_rematerialized",
-        initial_state=diamond.with_all_eager(),
+        initial_spec=diamond.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A", "B", "C", "D"]))
         .evaluate_tick()
         .assert_requested_runs()
@@ -253,7 +253,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="diamond_root_and_sink_rematerialized",
-        initial_state=diamond.with_all_eager(),
+        initial_spec=diamond.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A", "B", "C", "D"]))
         .evaluate_tick()
         .assert_requested_runs()
@@ -263,14 +263,14 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="parents_materialized_separate_runs",
-        initial_state=three_assets_in_sequence.with_all_eager(),
+        initial_spec=three_assets_in_sequence.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A"]), run_request(["B"]))
         .evaluate_tick()
         .assert_requested_runs(run_request(["C"])),
     ),
     AssetDaemonScenario(
         id="parent_materialized_twice",
-        initial_state=two_assets_in_sequence.with_all_eager(),
+        initial_spec=two_assets_in_sequence.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A"]), run_request(["A"]))
         .evaluate_tick()
         .assert_requested_runs(run_request(["B"]))
@@ -279,7 +279,7 @@ basic_scenarios = [
     ),
     AssetDaemonScenario(
         id="parent_rematerialized_twice",
-        initial_state=two_assets_in_sequence.with_all_eager(),
+        initial_spec=two_assets_in_sequence.with_all_eager(),
         execution_fn=lambda state: state.with_runs(run_request(["A", "B"]))
         .evaluate_tick()
         .assert_requested_runs()
