+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/updated_scenarios/custom_condition_scenarios.py
@@ -11,7 +11,7 @@ from .asset_daemon_scenario_states import one_asset, two_assets_in_sequence
 custom_condition_scenarios = [
     AssetDaemonScenario(
         id="funky_custom_condition_scenario",
-        initial_state=one_asset.with_asset_properties(
+        initial_spec=one_asset.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.from_asset_condition(
                 # intentionally funky condition
                 # not ((not missing) | (parent_updated)) ->
@@ -28,7 +28,7 @@ custom_condition_scenarios = [
     ),
     AssetDaemonScenario(
         id="funky_custom_condition_static_constructor_scenario",
-        initial_state=one_asset.with_asset_properties(
+        initial_spec=one_asset.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.from_asset_condition(
                 # same as above, but with static constructor
                 ~(~AssetCondition.missing() & AssetCondition.parent_newer())
@@ -40,7 +40,7 @@ custom_condition_scenarios = [
     ),
     AssetDaemonScenario(
         id="parent_newer_and_not_updated_since_cron_scenario",
-        initial_state=two_assets_in_sequence.with_asset_properties(
+        initial_spec=two_assets_in_sequence.with_asset_properties(
             "B",
             auto_materialize_policy=AutoMaterializePolicy.from_asset_condition(
                 ~AssetCondition.updated_since_cron("@daily") & AssetCondition.parent_newer()
