+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/updated_scenarios/freshness_policy_scenarios.py
@@ -1,9 +1,10 @@
 from dagster import AutoMaterializePolicy, FreshnessPolicy
 from dagster._core.definitions.asset_spec import AssetSpec
 
+from dagster_tests.definitions_tests.auto_materialize_tests.scenario_state import ScenarioSpec
+
 from ..asset_daemon_scenario import (
     AssetDaemonScenario,
-    AssetDaemonScenarioState,
     day_partition_key,
 )
 from ..base_scenario import run_request
@@ -19,30 +20,28 @@ from .asset_daemon_scenario_states import (
 
 freshness_30m = FreshnessPolicy(maximum_lag_minutes=30)
 freshness_60m = FreshnessPolicy(maximum_lag_minutes=60)
-extended_diamond = AssetDaemonScenarioState(
+extended_diamond = ScenarioSpec(
     asset_specs=[*diamond.asset_specs, AssetSpec("E", deps=["C"]), AssetSpec("F", deps=["D"])]
 )
 
 freshness_policy_scenarios = [
     AssetDaemonScenario(
         id="one_asset_lazy_never_materialized",
-        initial_state=one_asset.with_asset_properties(
+        initial_spec=one_asset.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy()
         ),
         execution_fn=lambda state: state.evaluate_tick().assert_requested_runs(),
     ),
     AssetDaemonScenario(
         id="one_asset_lazy_never_materialized_nothing_dep",
-        initial_state=AssetDaemonScenarioState(
-            asset_specs=[AssetSpec("B", deps=["A"])]
-        ).with_asset_properties(
+        initial_spec=ScenarioSpec(asset_specs=[AssetSpec("B", deps=["A"])]).with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(), freshness_policy=freshness_30m
         ),
         execution_fn=lambda state: state.evaluate_tick().assert_requested_runs(run_request("B")),
     ),
     AssetDaemonScenario(
         id="one_asset_lazy_with_freshness_policy_never_materialized",
-        initial_state=one_asset.with_asset_properties(
+        initial_spec=one_asset.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(),
             freshness_policy=FreshnessPolicy(maximum_lag_minutes=10),
         ),
@@ -50,7 +49,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="two_assets_eager_with_freshness_policies",
-        initial_state=two_assets_in_sequence.with_asset_properties(
+        initial_spec=two_assets_in_sequence.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.eager(),
             freshness_policy=FreshnessPolicy(maximum_lag_minutes=1000),
         ),
@@ -62,7 +61,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="one_asset_depends_on_two_lazy",
-        initial_state=one_asset_depends_on_two.with_asset_properties(
+        initial_spec=one_asset_depends_on_two.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(),
             freshness_policy=freshness_30m,
         ),
@@ -90,7 +89,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="diamond_lazy_basic",
-        initial_state=diamond.with_asset_properties(
+        initial_spec=diamond.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(),
             freshness_policy=freshness_30m,
         ),
@@ -119,7 +118,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="diamond_lazy_half_run_stale",
-        initial_state=diamond.with_asset_properties(
+        initial_spec=diamond.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(),
             freshness_policy=freshness_30m,
         ),
@@ -131,7 +130,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="diamond_lazy_half_run_and_failures",
-        initial_state=diamond.with_asset_properties(
+        initial_spec=diamond.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(),
             freshness_policy=freshness_30m,
         ),
@@ -178,7 +177,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="diamond_lazy_root_unselected",
-        initial_state=diamond.with_asset_properties(
+        initial_spec=diamond.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(),
             freshness_policy=freshness_30m,
         ).with_asset_properties(keys=["A"], auto_materialize_policy=None),
@@ -192,7 +191,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="extended_diamond_lazy",
-        initial_state=extended_diamond.with_asset_properties(
+        initial_spec=extended_diamond.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(),
         )
         .with_asset_properties(keys=["E"], freshness_policy=freshness_30m)
@@ -219,7 +218,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="two_assets_depend_on_one_lazy_with_cron_freshness",
-        initial_state=two_assets_depend_on_one.with_asset_properties(
+        initial_spec=two_assets_depend_on_one.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(),
         )
         .with_asset_properties(keys=["B"], freshness_policy=freshness_30m)
@@ -241,7 +240,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="extended_diamond_with_source_lazy",
-        initial_state=extended_diamond.with_asset_properties(
+        initial_spec=extended_diamond.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.lazy(),
         )
         .with_asset_properties(keys=["A"], deps=["source"])
@@ -255,7 +254,7 @@ freshness_policy_scenarios = [
     ),
     AssetDaemonScenario(
         id="daily_to_unpartitioned_lazy",
-        initial_state=two_assets_in_sequence.with_asset_properties(
+        initial_spec=two_assets_in_sequence.with_asset_properties(
             "A", partitions_def=daily_partitions_def
         )
         .with_asset_properties(
