+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/updated_scenarios/latest_materialization_run_tag_scenarios.py
@@ -24,7 +24,7 @@ filter_latest_run_tag_key_policy = (
 latest_materialization_run_tag_scenarios = [
     AssetDaemonScenario(
         id="latest_parent_materialization_has_required_tag",
-        initial_state=two_assets_in_sequence.with_asset_properties(
+        initial_spec=two_assets_in_sequence.with_asset_properties(
             auto_materialize_policy=filter_latest_run_tag_key_policy
         ),
         execution_fn=lambda state: state.with_runs(
@@ -35,7 +35,7 @@ latest_materialization_run_tag_scenarios = [
     ),
     AssetDaemonScenario(
         id="latest_parent_materialization_missing_required_tag",
-        initial_state=two_assets_in_sequence.with_asset_properties(
+        initial_spec=two_assets_in_sequence.with_asset_properties(
             auto_materialize_policy=filter_latest_run_tag_key_policy
         ),
         execution_fn=lambda state: state.with_runs(run_request(["A", "B"]), run_request(["A"]))
@@ -44,7 +44,7 @@ latest_materialization_run_tag_scenarios = [
     ),
     AssetDaemonScenario(
         id="latest_parent_materialization_missing_one_of_required_tags",
-        initial_state=two_assets_in_sequence.with_asset_properties(
+        initial_spec=two_assets_in_sequence.with_asset_properties(
             auto_materialize_policy=AutoMaterializePolicy.eager()
             .without_rules(AutoMaterializeRule.materialize_on_parent_updated())
             .with_rules(
@@ -66,7 +66,7 @@ latest_materialization_run_tag_scenarios = [
     ),
     AssetDaemonScenario(
         id="earlier_parent_materialization_missing_required_tag",
-        initial_state=two_assets_in_sequence.with_asset_properties(
+        initial_spec=two_assets_in_sequence.with_asset_properties(
             auto_materialize_policy=filter_latest_run_tag_key_policy
         ),
         execution_fn=lambda state: state.with_runs(
@@ -79,7 +79,7 @@ latest_materialization_run_tag_scenarios = [
     ),
     AssetDaemonScenario(
         id="only_updated_parent_missing_required_tag",
-        initial_state=one_asset_depends_on_two.with_asset_properties(
+        initial_spec=one_asset_depends_on_two.with_asset_properties(
             auto_materialize_policy=filter_latest_run_tag_key_policy
         ),
         execution_fn=lambda state: state.with_runs(
@@ -91,7 +91,7 @@ latest_materialization_run_tag_scenarios = [
     ),
     AssetDaemonScenario(
         id="one_updated_parent_has_required_tag_but_other_does_not",
-        initial_state=one_asset_depends_on_two.with_asset_properties(
+        initial_spec=one_asset_depends_on_two.with_asset_properties(
             auto_materialize_policy=filter_latest_run_tag_key_policy
         ),
         execution_fn=lambda state: state.with_runs(
@@ -104,7 +104,7 @@ latest_materialization_run_tag_scenarios = [
     ),
     AssetDaemonScenario(
         id="latest_materialization_missing_required_tag_but_will_update",
-        initial_state=three_assets_in_sequence.with_asset_properties(
+        initial_spec=three_assets_in_sequence.with_asset_properties(
             auto_materialize_policy=filter_latest_run_tag_key_policy
         ),
         execution_fn=lambda state: state.with_runs(
@@ -119,7 +119,7 @@ latest_materialization_run_tag_scenarios = [
         # this ensures that updates that happened prior to the cursor of the current tick
         # get ignored if they're missing the required keys
         id="latest_materialization_missing_required_tag_previous_tick",
-        initial_state=one_asset_depends_on_two.with_asset_properties(
+        initial_spec=one_asset_depends_on_two.with_asset_properties(
             auto_materialize_policy=filter_latest_run_tag_key_policy
         ),
         execution_fn=lambda state: state.with_runs(
